clarified prompts added alerts after each action
"Are you ready to begin" is clearer and more natural than "Do you wish to add or remove a plant".

Same with "Do you wish to add or remove another plant?"  "Do you wish to continue" is better.

I tried to write a summary alert in the loop and at the end of the addRemove function, but it kept causing an error that I couldn't catch, so instead, I added alerts after each action.  Success so far!