added if statements
Now it is grammatically correct depending on the number of plants.  YES. Getting this  to work was tough for me.  I thought it would be the same solution as in bottles of beer, but it was different because the indoor and outdoor plant words must be independent of each other.  It took me a long time to figure that out. I hope the neighbors are not disturbed by my loud cheering and clapping!